requirements.txt
Django>=3.2,<3.3
psycopg2>=2.8.6,<2.9
uWSGI>=2.0.19.1,<2.1
django-crispy-forms==1.14.0
canvasapi==3.0.0
pillow==10.2.0
django-allauth==0.60.1
numpy==1.26.2
numpy==1.26.2
cryptography==42.0.2
qrcode==7.4.2